// pslpwcl/src/app/modules/pslpapi/model/utente.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Cpi } from './cpi';
import { LavSapSez6 } from './lavSapSez6';
import { Ruolo } from './ruolo';

export interface Utente { 
    cfUtente?: string;
    cfUtenteOld?: string;
    codUserAggiorn?: string;
    codUserInserim?: string;
    cognome?: string;
    cpiComp?: Cpi;
    dAggiorn?: Date;
    dInserim?: Date;
    idSilLavAnagrafica?: number;
    idUtente?: number;
    identificativoSap?: string;
    lavSapSez6?: Array<LavSapSez6>;
    nome?: string;
    ruoli?: Array<Ruolo>;
}